+++ src/tests/components/IconList.test.ts
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/svelte';
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import IconList from '../../src/components/icons/IconList.svelte';
-import * as _icon from '../../src/models/Icon';
+import IconList from '../../components/icons/IconList.svelte';
+import * as _icon from '../../models/Icon';
 
 const iconsPromise = vi.hoisted(() => import('../mocks/icon.store.mock'));
