+++ src/tests/components/Toolbar.test.ts
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/svelte';
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import Toolbar from '../../src/components/toolbar/Toolbar.svelte';
+import Toolbar from '../../components/toolbar/Toolbar.svelte';
 import _ from 'lodash';
 
 const iconCustomizationsStorePromise = vi.hoisted(() => import('../mocks/icon-customizations.store.mock'));
